feat(ci): run workflows on merge_group (#8)
## Description

We have set `tests-pass` and `conventional-pr-title` as required for the
PRs to be merged. With the new merge queue activated, a new
`merge_group` step is introduced which requires again for these
workflows to run, as github doesn't allow switching that off for now, we
enable those workflows again with this PR for `merge_group`.